OSX: fix warnings/errors thrown by clang on OS X
gnrc_ipv6_nc, gnrc_sixlowpan, gnrc_netdev2_eth, gnrc_ipv6, xbee, sixlowpan,
sc_gnrc_6ctx

+++ sys/net/gnrc/network_layer/sixlowpan/gnrc_sixlowpan.c
@@ -303,7 +303,7 @@ static void _send(gnrc_pktsnip_t *pkt)
     }
     else {
         DEBUG("6lo: packet too big (%u > %" PRIu16 ")\n",
-              (unsigned int)datagram_size, SIXLOWPAN_FRAG_MAX_LEN);
+              (unsigned int)datagram_size, (uint16_t)SIXLOWPAN_FRAG_MAX_LEN);
         gnrc_pktbuf_release(pkt2);
     }
 #else
